@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Chubbyphp\Tests\Container\Unit;
 
 use Chubbyphp\Container\Parameter;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -14,9 +15,7 @@ use PHPUnit\Framework\TestCase;
  */
 final class ParameterTest extends TestCase
 {
-    /**
-     * @dataProvider provideInvokeCases
-     */
+    #[DataProvider('provideInvokeCases')]
     public function testInvoke(mixed $data): void
     {
         $parameter = new Parameter($data);
